requirements.txt
gspread
vk_api
smtplib
email
fuzzywuzzy
python-dotenv